+++ samples/dnn/layer_example_i16f32.c
@@ -75,28 +75,28 @@ typedef struct {
 
 LIBXS_INLINE void zero_buf_int16(short* buf, size_t size) {
   int i;
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = 0;
   }
 }
 
 LIBXS_INLINE void zero_buf_int32(int* buf, size_t size) {
   int i;
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = 0;
   }
 }
 
 LIBXS_INLINE void copy_buf_int16(short* src, short* dst, size_t size) {
   int i;
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     dst[i] = src[i];
   }
 }
 
 LIBXS_INLINE void zero_buf_f32(float* buf, size_t size) {
   int i;
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = 0;
   }
 }
@@ -105,7 +105,7 @@ LIBXS_INLINE void init_buf_int16(short* buf, size_t size, int initPos, int initO
 {
   int i;
   zero_buf_int16(buf, size);
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = (short)((initOne != 0) ? 1 : ((initPos != 0) ? (rand()%7) : (rand()%7-3)));
   }
 }
@@ -114,7 +114,7 @@ LIBXS_INLINE void init_buf_int32(int* buf, size_t size, int initPos, int initOne
 {
   int i;
   zero_buf_int32(buf, size);
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = (int)((initOne != 0) ? 1 : ((initPos != 0) ? (rand()%7) : (rand()%7-3)));
   }
 }
